overlay|refactor: Rename runChezmoi.ts to run-chezmoi.ts
Renames `packages/overlay/src/chezmoi/runChezmoi.ts` to `run-chezmoi.ts`, the kebab-case name the file-naming rule gives a module whose exports are peers with no single main export, and renames its test file to match.

`runChezmoiCaptured` and `runChezmoiStreamed` keep their names. Only the module path changes, across the ten importers and the renamed test, which imports its own subject.
