+++ packages/overlay/src/modes/create.ts
@@ -2,8 +2,8 @@ import path from 'node:path';
 
 import { parseStatus } from '../chezmoi/parseStatus.ts';
 import { readStatus } from '../chezmoi/readStatus.ts';
-import type { ChezmoiContext } from '../chezmoi/runChezmoi.ts';
-import { runChezmoiStreamed } from '../chezmoi/runChezmoi.ts';
+import type { ChezmoiContext } from '../chezmoi/run-chezmoi.ts';
+import { runChezmoiStreamed } from '../chezmoi/run-chezmoi.ts';
 import { countOutcome, partitionStatus } from './entry-outcomes.ts';
 import type { OverlayResult } from './types.ts';
