+++ packages/overlay/src/modes/force.ts
@@ -1,7 +1,7 @@
 import { parseStatus } from '../chezmoi/parseStatus.ts';
 import { readStatus } from '../chezmoi/readStatus.ts';
-import type { ChezmoiContext } from '../chezmoi/runChezmoi.ts';
-import { runChezmoiStreamed } from '../chezmoi/runChezmoi.ts';
+import type { ChezmoiContext } from '../chezmoi/run-chezmoi.ts';
+import { runChezmoiStreamed } from '../chezmoi/run-chezmoi.ts';
 import { countOutcome, partitionStatus } from './entry-outcomes.ts';
 import type { OverlayResult } from './types.ts';
