+++ packages/overlay/src/modes/verify.ts
@@ -1,6 +1,6 @@
 import { parseStatus } from '../chezmoi/parseStatus.ts';
 import { readStatus } from '../chezmoi/readStatus.ts';
-import type { ChezmoiContext } from '../chezmoi/runChezmoi.ts';
+import type { ChezmoiContext } from '../chezmoi/run-chezmoi.ts';
 import { partitionStatus } from './entry-outcomes.ts';
 import type { OverlayResult } from './types.ts';
